Move entity deduplication to consolidateQueryResult
